+++ src/shapes/Object/Object.ts
@@ -477,7 +477,7 @@ export class FabricObject<
    * and refreshed at the next render
    * @type Array
    */
-  declare cacheProperties: string[];
+  static cacheProperties: string[] = cacheProperties;
 
   /**
    * a fabricObject that, without stroke define a clipping area with their shape. filled in black
@@ -995,7 +995,8 @@ export class FabricObject<
   }
 
   /**
-   * @private
+   * Handles setting values on the instance and handling internal side effects
+   * @protected
    * @param {String} key
    * @param {*} value
    */
@@ -1019,7 +1020,8 @@ export class FabricObject<
 
     if (isChanged) {
       const groupNeedsUpdate = this.group && this.group.isOnACache();
-      if (this.cacheProperties.includes(key)) {
+      // @ts-ignore TS and constructor issue
+      if (this.constructor.cacheProperties.includes(key)) {
         this.dirty = true;
         groupNeedsUpdate && this.group!.set('dirty', true);
       } else if (groupNeedsUpdate && this.stateProperties.includes(key)) {
@@ -1824,11 +1826,9 @@ export class FabricObject<
    * override if necessary to dispose artifacts such as `clipPath`
    */
   dispose() {
-    // todo verify this.
-    // runningAnimations is always truthy
-    if (runningAnimations) {
-      runningAnimations.cancelByTarget(this);
-    }
+    runningAnimations.cancelByTarget(this);
+    this.off();
+    this._set('canvas', undefined);
   }
 
   /**
@@ -1885,7 +1885,6 @@ export class FabricObject<
  * For inheritance reasons ( used in the superclass but static in the subclass )
  */
 Object.assign(FabricObject.prototype, {
-  cacheProperties,
   stateProperties,
   ...fabricObjectDefaultValues,
 });