+++ test/unit/textbox.js
@@ -98,7 +98,7 @@
     assert.equal(textbox.text, 'test');
     assert.equal(textbox.type, 'textbox');
     assert.deepEqual(textbox.styles, { });
-    assert.ok(textbox.cacheProperties.indexOf('width') > -1, 'width is in cacheProperties');
+    assert.ok(fabric.Textbox.cacheProperties.indexOf('width') > -1, 'width is in cacheProperties');
   });
 
   QUnit.test('toObject', function(assert) {
@@ -178,7 +178,9 @@
       assert.deepEqual(obj.styles[2], textbox.styles[2], 'styles match at line 2');
       assert.deepEqual(obj.styles[2][0], textbox.styles[2][0], 'styles match at index 0');
       assert.deepEqual(obj.styles[2][1], textbox.styles[2][1], 'styles match at index 1');
-      fabric.Textbox.fromObject(obj).then(function(obj2) {
+      const out = obj.toObject();
+      fabric.Textbox.fromObject(out).then(function(obj2) {
+        assert.notEqual(out.styles, obj2.styles, 'styles copy is a different object after initialization');
         assert.notEqual(obj.styles, obj2.styles, 'styles copy is a different object after initialization');
         assert.deepEqual(obj.styles, obj2.styles, 'styles copy is a different object after initialization');
         done();
@@ -530,8 +532,7 @@
     var text = 'aaa aaq ggg gg oee eee';
     var styles = {};
     for (var index = 0; index < text.length; index++) {
-      styles[index] = { fontSize: 4 };
-      
+      styles[index] = { fontSize: 4 };      
     }
     var textbox = new fabric.Textbox(text, {
       styles: { 0: styles },